+++ src/controller_rpc.c
@@ -1895,7 +1895,7 @@ rpc_get_device_config(clixon_handle h,
         case DT_CANDIDATE:
         case DT_ACTIONS:
             xroot1 = xpath_first(xn, nsc, "config");
-            if (clixon_xml2cbuf(cb, xroot1, 0, 0, NULL, -1, 0) < 0)
+            if (clixon_xml2cbuf1(cb, xroot1, 0, 0, NULL, -1, 0, 0) < 0)
                 goto done;
             break;
         case DT_SYNCED:
@@ -3044,10 +3044,10 @@ rpc_device_config_template_apply(clixon_handle h,
         }
         if ((xtc = xml_dup(xtmpl)) == NULL)
             goto done;
-        if ((ret = xml_bind_yang(h, xtc, YB_MODULE, yspec1, &xerr)) < 0)
+        if ((ret = xml_bind_yang(h, xtc, YB_MODULE, yspec1, 0, &xerr)) < 0)
             goto done;
         if (ret == 0){
-            if (clixon_xml2cbuf(cbret, xerr, 0, 0, NULL, -1, 0) < 0)
+            if (clixon_xml2cbuf1(cbret, xerr, 0, 0, NULL, -1, 0, 0) < 0)
                 goto done;
             goto ok;
         }
@@ -3588,7 +3588,7 @@ controller_edit_config(clixon_handle h,
     if ((xc = xpath_first(xe, nsc, "%s", NETCONF_INPUT_CONFIG)) == NULL){
         goto ok;
     }
-    if ((ret = xml_bind_yang(h, xc, YB_MODULE, yspec, NULL)) < 0)
+    if ((ret = xml_bind_yang(h, xc, YB_MODULE, yspec, 0, NULL)) < 0)
         goto done;
     if (ret == 0)
         goto ok;
@@ -3602,7 +3602,7 @@ controller_edit_config(clixon_handle h,
     }
     if ((xserv = xml_find_type(xconfig, NULL, "services", CX_ELMNT)) == NULL)
         goto ok;
-    if ((ret = xml_bind_yang0(h, xserv, YB_MODULE, yspec, NULL)) < 0)
+    if ((ret = xml_bind_yang0(h, xserv, YB_MODULE, yspec, 0, NULL)) < 0)
         goto done;
     if (ret == 0)
         goto ok;